rename statusbar
