+++ src/packages/2StatusBar/index.ts
@@ -1,3 +1,3 @@
-export * from './src/2Components/2StatusBar/2StatusBar';
-export * from './src/2Components/2StatusItem/2StatusItem';
+export * from './src/2Components/StatusBar/StatusBar';
+export * from './src/2Components/StatusItem/StatusItem';
 export * from './src/Utils';
